requirements.txt
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
APScheduler==3.10.4
attrs==22.1.0
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
commonroad-drivability-checker==2022.1
commonroad-io==2022.1
-e git+https://gitlab.lrz.de/tum-cps/commonroad-route-planner.git@271e13b51cce8ea5494d9f13a611f46a57207915#egg=commonroad_route_planner
commonroad-vehicle-models==2.0.0
contourpy==1.0.5
coverage==6.5.0
cycler==0.11.0
exceptiongroup==1.0.4
Flask==2.2.5
Flask-APScheduler==1.13.0
Flask-Cors==4.0.0
Flask-HTTPAuth==4.7.0
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
fonttools==4.38.0
greenlet==2.0.2
idna==3.4
inflect==7.0.0
iniconfig==1.1.1
is-safe-url==1.0
iso3166==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.1
Mako==1.3.0
mariadb==1.1.7
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.29.0
matplotlib==3.3.4
mpld3==0.5.8
networkx==2.8.7
numpy==1.23.4
omegaconf==2.2.3
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
Polygon3==3.0.9.1
protobuf==3.20.1
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
pytest-docker==3.1.1
pytest-flask==1.3.0
pytest-mock==3.10.0
pytest-timeout==2.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0
requests==2.28.2
Rtree==1.0.1
scipy==1.9.3
Shapely==1.8.5.post1
six==1.16.0
SQLAlchemy==2.0.20
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==1.26.15
Werkzeug==2.2.2
